Also add a util.rs
Created using spr 1.3.6-beta.1
